Packages refactoring
