+++ .flutter-plugins-dependencies
@@ -1,1 +1,1 @@
-{"info":"This is a generated file; do not edit or check into version control.","plugins":{"ios":[{"name":"path_provider_foundation","path":"/Users/alekseizubankov/.pub-cache/hosted/pub.dev/path_provider_foundation-2.2.4/","shared_darwin_source":true,"native_build":true,"dependencies":[]},{"name":"sqflite","path":"/Users/alekseizubankov/.pub-cache/hosted/pub.dev/sqflite-2.3.0/","native_build":true,"dependencies":[]},{"name":"yandex_mapkit","path":"/Users/alekseizubankov/.pub-cache/git/yandex_mapkit-4341a01b474fb2b71a29f9dc0475fbe0fc4a1c56/","native_build":true,"dependencies":[]}],"android":[{"name":"flutter_plugin_android_lifecycle","path":"/Users/alekseizubankov/.pub-cache/hosted/pub.dev/flutter_plugin_android_lifecycle-2.0.15/","native_build":true,"dependencies":[]},{"name":"path_provider_android","path":"/Users/alekseizubankov/.pub-cache/hosted/pub.dev/path_provider_android-2.0.27/","native_build":true,"dependencies":[]},{"name":"sqflite","path":"/Users/alekseizubankov/.pub-cache/hosted/pub.dev/sqflite-2.3.0/","native_build":true,"dependencies":[]},{"name":"yandex_mapkit","path":"/Users/alekseizubankov/.pub-cache/git/yandex_mapkit-4341a01b474fb2b71a29f9dc0475fbe0fc4a1c56/","native_build":true,"dependencies":["flutter_plugin_android_lifecycle"]}],"macos":[{"name":"path_provider_foundation","path":"/Users/alekseizubankov/.pub-cache/hosted/pub.dev/path_provider_foundation-2.2.4/","shared_darwin_source":true,"native_build":true,"dependencies":[]},{"name":"sqflite","path":"/Users/alekseizubankov/.pub-cache/hosted/pub.dev/sqflite-2.3.0/","native_build":true,"dependencies":[]}],"linux":[{"name":"path_provider_linux","path":"/Users/alekseizubankov/.pub-cache/hosted/pub.dev/path_provider_linux-2.1.11/","native_build":false,"dependencies":[]}],"windows":[{"name":"path_provider_windows","path":"/Users/alekseizubankov/.pub-cache/hosted/pub.dev/path_provider_windows-2.1.7/","native_build":false,"dependencies":[]}],"web":[]},"dependencyGraph":[{"name":"flutter_plugin_android_lifecycle","dependencies":[]},{"name":"path_provider","dependencies":["path_provider_android","path_provider_foundation","path_provider_linux","path_provider_windows"]},{"name":"path_provider_android","dependencies":[]},{"name":"path_provider_foundation","dependencies":[]},{"name":"path_provider_linux","dependencies":[]},{"name":"path_provider_windows","dependencies":[]},{"name":"sqflite","dependencies":[]},{"name":"yandex_mapkit","dependencies":["flutter_plugin_android_lifecycle"]}],"date_created":"2023-08-09 10:17:56.244071","version":"3.10.6"}
+{"info":"This is a generated file; do not edit or check into version control.","plugins":{"ios":[{"name":"path_provider_foundation","path":"/Users/alekseizubankov/.pub-cache/hosted/pub.dev/path_provider_foundation-2.2.4/","shared_darwin_source":true,"native_build":true,"dependencies":[]},{"name":"sqflite","path":"/Users/alekseizubankov/.pub-cache/hosted/pub.dev/sqflite-2.3.0/","native_build":true,"dependencies":[]},{"name":"yandex_mapkit","path":"/Users/alekseizubankov/.pub-cache/git/yandex_mapkit-4341a01b474fb2b71a29f9dc0475fbe0fc4a1c56/","native_build":true,"dependencies":[]}],"android":[{"name":"flutter_plugin_android_lifecycle","path":"/Users/alekseizubankov/.pub-cache/hosted/pub.dev/flutter_plugin_android_lifecycle-2.0.15/","native_build":true,"dependencies":[]},{"name":"path_provider_android","path":"/Users/alekseizubankov/.pub-cache/hosted/pub.dev/path_provider_android-2.0.27/","native_build":true,"dependencies":[]},{"name":"sqflite","path":"/Users/alekseizubankov/.pub-cache/hosted/pub.dev/sqflite-2.3.0/","native_build":true,"dependencies":[]},{"name":"yandex_mapkit","path":"/Users/alekseizubankov/.pub-cache/git/yandex_mapkit-4341a01b474fb2b71a29f9dc0475fbe0fc4a1c56/","native_build":true,"dependencies":["flutter_plugin_android_lifecycle"]}],"macos":[{"name":"path_provider_foundation","path":"/Users/alekseizubankov/.pub-cache/hosted/pub.dev/path_provider_foundation-2.2.4/","shared_darwin_source":true,"native_build":true,"dependencies":[]},{"name":"sqflite","path":"/Users/alekseizubankov/.pub-cache/hosted/pub.dev/sqflite-2.3.0/","native_build":true,"dependencies":[]}],"linux":[{"name":"path_provider_linux","path":"/Users/alekseizubankov/.pub-cache/hosted/pub.dev/path_provider_linux-2.1.11/","native_build":false,"dependencies":[]}],"windows":[{"name":"path_provider_windows","path":"/Users/alekseizubankov/.pub-cache/hosted/pub.dev/path_provider_windows-2.1.7/","native_build":false,"dependencies":[]}],"web":[]},"dependencyGraph":[{"name":"flutter_plugin_android_lifecycle","dependencies":[]},{"name":"path_provider","dependencies":["path_provider_android","path_provider_foundation","path_provider_linux","path_provider_windows"]},{"name":"path_provider_android","dependencies":[]},{"name":"path_provider_foundation","dependencies":[]},{"name":"path_provider_linux","dependencies":[]},{"name":"path_provider_windows","dependencies":[]},{"name":"sqflite","dependencies":[]},{"name":"yandex_mapkit","dependencies":["flutter_plugin_android_lifecycle"]}],"date_created":"2023-08-29 13:29:59.170518","version":"3.10.6"}
