+++ lib/effective_map.dart
@@ -1,6 +1,5 @@
 library effective_map;
 
 export 'package:effective_map/src/models/models.dart';
-export 'package:effective_map/src/utils/utils.dart';
 
-export 'package:effective_map/src/view/effective_map.dart';
+export 'package:effective_map/src/maps/effective_map.dart';
