+++ lib/src/maps/i_map_state.dart
