+++ lib/src/maps/osm/map_controller/osm_effective_map_controller.dart
@@ -1,5 +1,6 @@
 import 'dart:math';
 
+import 'package:effective_map/src/maps/osm/utils/bbox_extension.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_map/flutter_map.dart';
 import 'package:flutter_map_animations/flutter_map_animations.dart';
@@ -7,8 +8,7 @@ import 'package:flutter_map_animations/flutter_map_animations.dart';
 import 'package:effective_map/src/models/bbox.dart';
 import 'package:effective_map/src/models/effective_latlng.dart';
 import 'package:effective_map/src/models/map_controller/effective_map_controller.dart';
-import 'package:effective_map/src/utils/flutter_map_extension.dart';
-import 'package:effective_map/src/utils/bbox_extension.dart';
+import 'package:effective_map/src/maps/osm/utils/flutter_map_extension.dart';
 
 const _maxCameraZoom = 19.0;
 const _interactivePolygonVisibilityThreshold = 17.3;
