+++ lib/src/maps/osm/utils/flutter_map_extension.dart
