+++ lib/src/maps/osm/view/osm_map.dart
@@ -11,7 +11,7 @@ import 'package:effective_map/src/models/effective_marker.dart';
 import 'package:effective_map/src/models/effective_network_tiles_provider.dart';
 import 'package:effective_map/src/models/effective_latlng.dart';
 import 'package:effective_map/src/models/map_controller/effective_map_controller.dart';
-import 'package:effective_map/src/models/map_controller/osm_effective_map_controller.dart';
+import 'package:effective_map/src/maps/osm/map_controller/osm_effective_map_controller.dart';
 import 'package:effective_map/src/models/map_object_with_geometry.dart';
 import 'package:effective_map/src/models/map_object_wrappers/circle_wrapper.dart';
 import 'package:effective_map/src/models/map_object_wrappers/map_object.dart';
@@ -19,14 +19,14 @@ import 'package:effective_map/src/models/map_object_wrappers/multi_polygon_wrapp
 import 'package:effective_map/src/models/map_object_wrappers/multi_polyline_wrapper.dart';
 import 'package:effective_map/src/models/map_object_wrappers/polygon_wrapper.dart';
 import 'package:effective_map/src/models/map_object_wrappers/polyline_wrapper.dart';
-import 'package:effective_map/i_map_state.dart';
-import 'package:effective_map/src/utils/bbox_extension.dart';
-import 'package:effective_map/src/utils/cached_tile_provider.dart';
-import 'package:effective_map/src/utils/flutter_map_extension.dart';
-import 'package:effective_map/src/utils/number_extractor.dart';
-import 'package:effective_map/src/view/widgets/any_map_object_layer.dart';
-import 'package:effective_map/src/view/widgets/cluster_widget.dart';
-import 'package:effective_map/src/view/widgets/user_location_layer.dart';
+import 'package:effective_map/src/maps/i_map_state.dart';
+import 'package:effective_map/src/maps/osm/utils/bbox_extension.dart';
+import 'package:effective_map/src/maps/osm/utils/cached_tile_provider.dart';
+import 'package:effective_map/src/maps/osm/utils/flutter_map_extension.dart';
+import 'package:effective_map/src/common/number_extractor.dart';
+import 'package:effective_map/src/maps/osm/view/widgets/any_map_object_layer.dart';
+import 'package:effective_map/src/maps/osm/view/widgets/cluster_widget.dart';
+import 'package:effective_map/src/maps/osm/view/widgets/user_location_layer.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_map/flutter_map.dart';
 import 'package:flutter_map/plugin_api.dart';
