+++ lib/src/maps/osm/view/widgets/user_location_layer.dart
@@ -4,7 +4,7 @@ import 'package:flutter_map/flutter_map.dart';
 import 'package:effective_map/src/common/constants.dart';
 import 'package:effective_map/src/common/package_colors.dart';
 import 'package:effective_map/src/models/effective_latlng.dart';
-import 'package:effective_map/src/utils/flutter_map_extension.dart';
+import 'package:effective_map/src/maps/osm/utils/flutter_map_extension.dart';
 
 const inactiveUserLocationShadow = <BoxShadow>[
   BoxShadow(
