+++ lib/src/maps/yandex/utils/geoobjects_canvas_painter.dart
