+++ lib/src/maps/yandex/utils/map_geometry_creator.dart
