+++ lib/src/maps/yandex/view/yandex_map.dart
@@ -9,16 +9,16 @@ import 'package:effective_map/src/models/effective_marker.dart';
 import 'package:effective_map/src/models/effective_network_tiles_provider.dart';
 import 'package:effective_map/src/models/effective_latlng.dart';
 import 'package:effective_map/src/models/map_controller/effective_map_controller.dart';
-import 'package:effective_map/src/models/map_controller/yandex_effective_map_controller.dart';
+import 'package:effective_map/src/maps/yandex/map_controller/yandex_effective_map_controller.dart';
 import 'package:effective_map/src/models/map_object_with_geometry.dart';
-import 'package:effective_map/i_map_state.dart';
-import 'package:effective_map/src/utils/bbox_extension.dart';
-import 'package:effective_map/src/utils/bounding_box_former.dart';
-import 'package:effective_map/src/utils/geoobjects_canvas_painter.dart';
-import 'package:effective_map/src/utils/map_geometry_creator.dart';
-import 'package:effective_map/src/utils/number_extractor.dart';
-import 'package:effective_map/src/utils/placemarks.dart';
-import 'package:effective_map/src/utils/yandex_map_extension.dart';
+import 'package:effective_map/src/maps/i_map_state.dart';
+import 'package:effective_map/src/maps/yandex/utils/bbox_extension.dart';
+import 'package:effective_map/src/maps/yandex/utils/bounding_box_former.dart';
+import 'package:effective_map/src/maps/yandex/utils/geoobjects_canvas_painter.dart';
+import 'package:effective_map/src/maps/yandex/utils/map_geometry_creator.dart';
+import 'package:effective_map/src/common/number_extractor.dart';
+import 'package:effective_map/src/maps/yandex/utils/placemarks.dart';
+import 'package:effective_map/src/maps/yandex/utils/yandex_map_extension.dart';
 import 'package:flutter/material.dart';
 import 'package:yandex_mapkit/yandex_mapkit.dart';
