--- lib/src/utils/utils.dart
@@ -1,8 +1,0 @@
-library utils;
-
-export 'bbox_extension.dart';
-
-export 'cached_tile_provider.dart';
-export 'flutter_map_extension.dart';
-
-export 'yandex_map_extension.dart';
